Added DetectionEfficiency program.
This program scans a single sky-coverage file,
and then extracts all the relevant observations from the
MPC catalog. It also computes for every know MPC object if
it was in the sky-coverage fields at the relevant epoch.
By combining all the relevant data, this program determines the
single-night efficiency as a function of the limiting magnitude
declared in the sky-coverage file.